[run all tests] updated matcher
